chore(server/middleware/auth): Add auth middleware
